requirements.txt
chromadb>=0.4.24
sentence-transformers==2.7.0
huggingface-hub<0.26.0
google-generativeai>=0.3.2
langchain>=0.2.0
langgraph>=0.2.0
langchain-community>=0.2.0
beautifulsoup4>=4.12.0
requests>=2.31.0
